+++ RELEASE_NOTES.md
@@ -2,3 +2,4 @@
 
 - Implement the [Cart API](https://developer.bigcommerce.com/api-reference/store-management/carts/cart/createacart).
 - Implement the [Cart Items API](https://developer.bigcommerce.com/api-reference/store-management/carts/cart-items/addcartlineitem)
+- Implement the [Cart Redirect URLS API](https://developer.bigcommerce.com/api-reference/store-management/carts/cart-redirect-urls/createcartredirecturl)